WIP reorganising cdsresponder
