+++ CDS/cdsresponder/lib/Network.rb
@@ -1,0 +1,7 @@
+require 'webrick'
+require 'HealthCheckServlet'
+
+def startup_networkreceiver
+  server = WEBrick::HTTPServer.new(:Port=>$options.port)
+  server.mount('/healthcheck',HealthCheckServlet)
+end
